+++ Euclid/CHANGELOG.md
@@ -1,3 +1,17 @@
+## [0.6.4](https://github.com/nicklockwood/Euclid/releases/tag/0.6.4) (2022-11-21)
+
+- Improved default alignment heuristic when extruding along non-planar paths
+- Added `Alignment` enum for controlling shape alignment when extruding along a path
+- Fixed bug in face normal calculation for open paths
+- Added scalar multiply/divide operators for `Rotation`
+
+## [0.6.3](https://github.com/nicklockwood/Euclid/releases/tag/0.6.3) (2022-11-17)
+
+- Fixed crash when extruding single-point path
+- Fixed `Quaternion.axis` when angle is zero
+- Disfavored `Angle.cos`/`sin`/`tan` overloads to improve ergonomics
+- Added `Polygon.area` property
+
 ## [0.6.2](https://github.com/nicklockwood/Euclid/releases/tag/0.6.2) (2022-11-05)
 
 - Added `Mesh.convexHull()` shape constructors